feat: added 410 response to GET /users

+++ internal/data/delete_user.go
@@ -1,17 +1,31 @@
 package data
 
-import "context"
-
-type queryDeleteUser struct{}
-
-func (q queryDeleteUser) text() string {
-	return `
-	DELETE FROM users
-	WHERE id = $1;
-	`
-}
+import (
+	"context"
+	"database/sql"
+)
 
 func (s *Storage) DeleteUser(ctx context.Context, userId string) (err error) {
-	_, err = s.queries[queryDeleteUser{}].ExecContext(ctx, userId)
-	return err
+	var tx *sql.Tx
+	tx, err = s.db.BeginTx(ctx, nil)
+
+	if err != nil {
+		return err
+	}
+
+	defer tx.Rollback()
+
+	_, err = tx.ExecContext(ctx, "DELETE FROM users WHERE id = $1;", userId)
+
+	if err != nil {
+		return err
+	}
+
+	_, err = tx.ExecContext(ctx, "INSERT INTO deleted_users (id) VALUES ($1);", userId)
+
+	if err != nil {
+		return err
+	}
+
+	return tx.Commit()
 }
